Refactor UI components and update directory structure
The commit reorganizes the UI components by moving them to a separate 'components' directory. It renames several files to better reflect their content and purpose. This change improves code organization and understanding, making it easier to maintain and develop further.
